+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
 
+### Added 
+ - metrics-docker-stats.rb: New option `--container-list-regex` to filter containers using Regex String (@parin921996).
+ 
 ## [3.1.1] - 2018
 ### Fixed
  - check-container-logs.rb: fix nil.gsub condition on empty log lines